svn path=/tags/KDE_3_1_3_RELEASE/kdepim/; revision=238664

+++ kpilot/lib/kpilotlink.cc
@@ -17,8 +17,8 @@
 **
 ** You should have received a copy of the GNU Lesser General Public License
 ** along with this program in a file called COPYING; if not, write to
-** the Free Software Foundation, Inc., 675 Mass Ave, Cambridge,
-** MA 02139, USA.
+** the Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
+** MA 02111-1307, USA.
 */
 
 /*
@@ -46,8 +46,9 @@ static const char *kpilotlink_id = "$Id$";
 #include <stdio.h>
 #include <unistd.h>
 #include <fcntl.h>
-#include <iostream.h>
 #include <errno.h>
+
+#include <iostream>
 
 #include <qdir.h>
 #include <qtimer.h>
@@ -98,6 +99,7 @@ KPilotDeviceLink::KPilotDeviceLink(QObject * parent, const char *name) :
 	fRetries(0),
 	fOpenTimer(0L),
 	fSocketNotifier(0L),
+	fSocketNotifierActive(false),
 	fPilotMasterSocket(-1),
 	fCurrentPilotSocket(-1)
 {
@@ -138,6 +140,7 @@ void KPilotDeviceLink::close()
 
 	KPILOT_DELETE(fOpenTimer);
 	KPILOT_DELETE(fSocketNotifier);
+	fSocketNotifierActive=false;
 #ifdef DEBUG
 	DEBUGDAEMON << fname
 		<< ": Closing sockets "
@@ -293,6 +296,9 @@ bool KPilotDeviceLink::open()
 			e = 0;
 			goto errInit;
 		}
+#ifdef DEBUG
+		DEBUGDAEMON << fname << ": Typing to open " << fPilotPath << endl;
+#endif
 
 #if PILOT_LINK_NUMBER < 10
 		fPilotMasterSocket = pi_socket(PI_AF_SLP,
@@ -344,6 +350,7 @@ bool KPilotDeviceLink::open()
 			QSocketNotifier::Read, this);
 		QObject::connect(fSocketNotifier, SIGNAL(activated(int)),
 			this, SLOT(acceptDevice()));
+		fSocketNotifierActive=true;
 		return true;
 	}
 	else
@@ -378,7 +385,7 @@ bool KPilotDeviceLink::open()
 errInit:
 	close();
 
-	if (msg.contains('%'))
+	if (msg.find('%'))
 	{
 		if (fPilotPath.isEmpty())
 		{
@@ -428,9 +435,17 @@ void KPilotDeviceLink::acceptDevice()
 
 	int ret;
 
+	if (!fSocketNotifierActive) 
+	{
+		kdWarning() << k_funcinfo << ": Accidentally in acceptDevice()"
+			<< endl;
+		return;
+	}
+
 	if (fSocketNotifier)
 	{
-		fSocketNotifier->setEnabled(false);
+		// fSocketNotifier->setEnabled(false);
+		fSocketNotifierActive=false;
 	}
 
 #ifdef DEBUG
@@ -447,8 +462,10 @@ void KPilotDeviceLink::acceptDevice()
 
 		kdWarning() << "pi_listen: " << s << endl;
 
+		// Presumably, strerror() returns things in
+		// local8Bit and not latin1.
 		emit logError(i18n("Can't listen on Pilot socket (%1)").
-			arg(s));
+			arg(QString::fromLocal8Bit(s)));
 
 		close();
 		return;
@@ -463,9 +480,11 @@ void KPilotDeviceLink::acceptDevice()
 
 		kdWarning() << "pi_accept: " << s << endl;
 
-		emit logError(i18n("Can't accept Pilot (%1)").arg(s));
+		emit logError(i18n("Can't accept Pilot (%1)")
+			.arg(QString::fromLocal8Bit(s)));
 
 		fStatus = PilotLinkError;
+		close();
 		return;
 	}
 
@@ -475,49 +494,13 @@ void KPilotDeviceLink::acceptDevice()
 		kdError() << k_funcinfo
 			<< ": Already connected or unable to connect!"
 			<< endl;
+		emit logError(TODO_I18N("Can't accept Pilot (%1)")
+			.arg(TODO_I18N("already connected")));
+		close();
 		return;
 	}
 
 	emit logProgress(QString::null, 30);
-
-	fPilotUser = new KPilotUser;
-
-	/* Ask the pilot who it is.  And see if it's who we think it is. */
-#ifdef DEBUG
-	DEBUGDAEMON << fname << ": Reading user info @"
-		<< (int) fPilotUser << endl;
-	DEBUGDAEMON << fname << ": Buffer @"
-		<< (int) fPilotUser->pilotUser() << endl;
-#endif
-
-	dlp_ReadUserInfo(fCurrentPilotSocket, fPilotUser->pilotUser());
-	fPilotUser->boundsCheck();
-
-#ifdef DEBUG
-	DEBUGDAEMON << fname
-		<< ": Read user name " << fPilotUser->getUserName() << endl;
-#endif
-
-	emit logProgress(i18n("Checking last PC..."), 60);
-
-	/* Tell user (via Pilot) that we are starting things up */
-	if ((ret=dlp_OpenConduit(fCurrentPilotSocket)) < 0)
-	{
-		kdWarning() << k_funcinfo
-			<< ": dlp_OpenConduit returned " << ret << endl;
-
-#if 0
-		fStatus = SyncDone;
-		emit logMessage(i18n
-			("Exiting on cancel. All data not restored."));
-		return;
-#endif
-		emit logError(i18n("Could not read user information from the Pilot. "
-			"Perhaps you have a password set on the device?"));
-	}
-	fStatus = AcceptedDevice;
-
-	emit logProgress(QString::null,70);
 
 	struct SysInfo sys_info;
 	if (dlp_ReadSysInfo(fCurrentPilotSocket,&sys_info) < 0)
@@ -540,6 +523,45 @@ void KPilotDeviceLink::acceptDevice()
 			<< endl;
 	}
 #endif
+	
+	emit logProgress(QString::null, 60);
+	fPilotUser = new KPilotUser;
+
+	/* Ask the pilot who it is.  And see if it's who we think it is. */
+#ifdef DEBUG
+	DEBUGDAEMON << fname << ": Reading user info @"
+		<< (int) fPilotUser << endl;
+	DEBUGDAEMON << fname << ": Buffer @"
+		<< (int) fPilotUser->pilotUser() << endl;
+#endif
+
+	dlp_ReadUserInfo(fCurrentPilotSocket, fPilotUser->pilotUser());
+	fPilotUser->boundsCheck();
+
+#ifdef DEBUG
+	DEBUGDAEMON << fname
+		<< ": Read user name " << fPilotUser->getUserName() << endl;
+#endif
+
+	emit logProgress(i18n("Checking last PC..."), 90);
+
+	/* Tell user (via Pilot) that we are starting things up */
+	if ((ret=dlp_OpenConduit(fCurrentPilotSocket)) < 0)
+	{
+		DEBUGDAEMON << k_funcinfo
+			<< ": dlp_OpenConduit returned " << ret << endl;
+
+#if 0
+		fStatus = SyncDone;
+		emit logMessage(i18n
+			("Exiting on cancel. All data not restored."));
+		return;
+#endif
+		emit logError(i18n("Could not read user information from the Pilot. "
+			"Perhaps you have a password set on the device?"));
+	}
+	fStatus = AcceptedDevice;
+
 
 	emit logProgress(QString::null, 100);
 	emit deviceReady();
@@ -552,7 +574,7 @@ void KPilotDeviceLink::tickle() const
 }
 
 
-int KPilotDeviceLink::installFiles(const QStringList & l)
+int KPilotDeviceLink::installFiles(const QStringList & l, const bool deleteFiles)
 {
 	FUNCTIONSETUP;
 
@@ -565,7 +587,7 @@ int KPilotDeviceLink::installFiles(const QStringList & l)
 		emit logProgress(QString::null,
 			(int) ((100.0 / l.count()) * (float) n));
 
-		if (installFile(*i))
+		if (installFile(*i, deleteFiles))
 			k++;
 		n++;
 	}
@@ -574,7 +596,7 @@ int KPilotDeviceLink::installFiles(const QStringList & l)
 	return k;
 }
 
-bool KPilotDeviceLink::installFile(const QString & f)
+bool KPilotDeviceLink::installFile(const QString & f, const bool deleteFile)
 {
 	FUNCTIONSETUP;
 
@@ -610,13 +632,13 @@ bool KPilotDeviceLink::installFile(const QString & f)
 	}
 
 	pi_file_close(pf);
-	QFile::remove(f);
+	if (deleteFile) QFile::remove(f);
 
 	return true;
 }
 
 
-void KPilotDeviceLink::addSyncLogEntry(const QString & entry, bool suppress)
+void KPilotDeviceLink::addSyncLogEntry(const QString & entry, bool log)
 {
 	FUNCTIONSETUP;
 
@@ -628,7 +650,7 @@ void KPilotDeviceLink::addSyncLogEntry(const QString & entry, bool suppress)
 	
 	dlp_AddSyncLogEntry(fCurrentPilotSocket,
 		const_cast < char *>(t.latin1()));
-	if (!suppress)
+	if (log)
 	{
 		emit logMessage(entry);
 	}
@@ -645,49 +667,49 @@ QString KPilotDeviceLink::deviceTypeString(int i) const
 	switch (i)
 	{
 	case None:
-		return QString("None");
+		return QString::fromLatin1("None");
 	case Serial:
-		return QString("Serial");
+		return QString::fromLatin1("Serial");
 	case OldStyleUSB:
-		return QString("OldStyleUSB");
+		return QString::fromLatin1("OldStyleUSB");
 	case DevFSUSB:
-		return QString("DevFSUSB");
+		return QString::fromLatin1("DevFSUSB");
 	default:
-		return QString("<unknown>");
+		return QString::fromLatin1("<unknown>");
 	}
 }
 
 QString KPilotDeviceLink::statusString() const
 {
 	FUNCTIONSETUP;
-	QString s("KPilotDeviceLink=");
+	QString s = QString::fromLatin1("KPilotDeviceLink=");
 
 
 	switch (fStatus)
 	{
 	case Init:
-		s.append("Init");
+		s.append(QString::fromLatin1("Init"));
 		break;
 	case WaitingForDevice:
-		s.append("WaitingForDevice");
+		s.append(QString::fromLatin1("WaitingForDevice"));
 		break;
 	case FoundDevice:
-		s.append("FoundDevice");
+		s.append(QString::fromLatin1("FoundDevice"));
 		break;
 	case CreatedSocket:
-		s.append("CreatedSocket");
+		s.append(QString::fromLatin1("CreatedSocket"));
 		break;
 	case DeviceOpen:
-		s.append("DeviceOpen");
+		s.append(QString::fromLatin1("DeviceOpen"));
 		break;
 	case AcceptedDevice:
-		s.append("AcceptedDevice");
+		s.append(QString::fromLatin1("AcceptedDevice"));
 		break;
 	case SyncDone:
-		s.append("SyncDone");
+		s.append(QString::fromLatin1("SyncDone"));
 		break;
 	case PilotLinkError:
-		s.append("PilotLinkError");
+		s.append(QString::fromLatin1("PilotLinkError"));
 		break;
 	}
 
@@ -715,10 +737,12 @@ int KPilotDeviceLink::getNextDatabase(int index,struct DBInfo *dbinfo)
 }
 
 // Find a database with the given name. Info about the DB is stored into dbinfo (e.g. to be used later on with retrieveDatabase).
-int KPilotDeviceLink::findDatabase(char*name, struct DBInfo*dbinfo) 
+int KPilotDeviceLink::findDatabase(const char *name, struct DBInfo *dbinfo,
+	int index, long type, long creator) 
 {
 	FUNCTIONSETUP;
-	return dlp_FindDBInfo(pilotSocket(), 0, 0, name, 0, 0, dbinfo);
+	return dlp_FindDBInfo(pilotSocket(), 0, index, 
+		const_cast<char *>(name), type, creator, dbinfo);
 }
 
 bool KPilotDeviceLink::retrieveDatabase(const QString &fullBackupName, 
@@ -798,11 +822,15 @@ unsigned long KPilotDeviceLink::minorVersion() const
 	return (((rom >> 20) & 0xf) * 10)+ ((rom >> 16) & 0xf);
 }
 
+/* static */ const int KPilotDeviceLink::messagesType=
+	(int)OpenFailMessage ;
+	
 void KPilotDeviceLink::shouldPrint(int m,const QString &s)
 {
 	if (!(messages & m))
 	{
-		emit logError(s);
+		if (messagesType & m) { emit logError(s); }
+		else { emit logMessage(s); }
 		messages |= (m & messagesMask);
 	}
 }
@@ -821,123 +849,3 @@ bool operator < (const db & a, const db & b) {
 
 	return a.maxblock < b.maxblock;
 }
-
-// $Log$
-// Revision 1.25  2002/08/31 22:35:46  mhunter
-// CVS_SILENT Corrected typographical errors
-//
-// Revision 1.24  2002/08/30 22:24:55  adridg
-// - Improved logging, connected the right signals now
-// - Try to handle dlp_ReadUserInfo failures sensibly
-// - Trying to sort out failures reading the database list.
-//
-// Revision 1.23  2002/08/24 21:27:32  adridg
-// Lots of small stuff to remove warnings
-//
-// Revision 1.22  2002/08/23 22:03:21  adridg
-// See ChangeLog - exec() becomes bool, debugging added
-//
-// Revision 1.21  2002/08/21 17:20:47  adridg
-// Detect and complain about common permissions errors
-//
-// Revision 1.20  2002/08/20 21:18:31  adridg
-// License change in lib/ to allow plugins -- which use the interfaces and
-// definitions in lib/ -- to use non-GPL'ed libraries, in particular to
-// allow the use of libmal which is MPL.
-//
-// Revision 1.19  2002/08/12 13:23:03  kainhofe
-// used long instead of unsigned long for dpl_ReadFeature. Fixed.
-//
-// Revision 1.18  2002/07/30 17:17:48  kainhofe
-// added majorVersion, minorVersion and ROMversion functions
-//
-// Revision 1.17  2002/07/25 22:11:22  kainhofe
-// time sync conduit
-//
-// Revision 1.16  2002/07/25 19:02:20  kainhofe
-// Added functions to get/set the time on the handheld
-//
-// Revision 1.15  2002/07/20 22:08:19  mhunter
-// Hot-Sync -> HotSync
-//
-// Revision 1.14  2002/07/05 00:15:22  kainhofe
-// Added KPilotDeviceLink::tickle(), Changelog update, compile fixes
-//
-// Revision 1.13  2002/05/18 23:28:19  adridg
-// Compile fixes
-//
-// Revision 1.12  2002/05/15 20:16:20  kainhofe
-// Wrongfully had a constant where 0 needed to be...
-//
-// Revision 1.11  2002/05/14 22:57:40  adridg
-// Merge from _BRANCH
-//
-// Revision 1.10  2002/05/03 17:21:51  kainhofe
-// Added a method findDatabase to KPilotDeviceLink to look up a single db on the palm
-//
-// Revision 1.9  2002/04/20 13:03:31  binner
-// CVS_SILENT Capitalisation fixes.
-//
-// Revision 1.8.2.3  2002/05/01 21:37:30  kainhofe
-// Added a function findDatabase to get the DBInfo for a given database
-//
-// Revision 1.8.2.2  2002/04/09 21:51:50  adridg
-// Extra debugging, pilot-link 0.10.1 still needs workaround
-//
-// Revision 1.8.2.1  2002/04/04 20:28:28  adridg
-// Fixing undefined-symbol crash in vcal. Fixed FD leak. Compile fixes
-// when using PILOT_VERSION. kpilotTest defaults to list, like the options
-// promise. Always do old-style USB sync (also works with serial devices)
-// and runs conduits only for HotSync. KPilot now as it should have been
-// for the 3.0 release.
-//
-// Revision 1.8  2002/02/10 22:21:33  adridg
-// Handle pilot-link 0.10.1; spit 'n polish; m505 now supported?
-//
-// Revision 1.7  2002/02/02 11:46:03  adridg
-// Abstracting away pilot-link stuff
-//
-// Revision 1.6  2002/01/25 21:43:13  adridg
-// ToolTips->WhatsThis where appropriate; vcal conduit discombobulated - it doesn't eat the .ics file anymore, but sync is limited; abstracted away more pilot-link
-//
-// Revision 1.5  2002/01/21 23:14:03  adridg
-// Old code removed; extra abstractions added; utility extended
-//
-// Revision 1.4  2002/01/18 16:28:57  adridg
-// CVS_SILENT: Less compile warnings
-//
-// Revision 1.3  2001/12/29 15:45:28  adridg
-// Lots of little changes for the syncstack
-//
-// Revision 1.2  2001/10/08 22:25:41  adridg
-// Moved to libkpilot and lib-based conduits
-//
-// Revision 1.1  2001/10/08 21:56:02  adridg
-// Start of making a separate KPilot lib
-//
-// Revision 1.60  2001/09/30 19:51:56  adridg
-// Some last-minute layout, compile, and __FUNCTION__ (for Tru64) changes.
-//
-// Revision 1.59  2001/09/29 16:26:18  adridg
-// The big layout change
-//
-// Revision 1.58  2001/09/24 22:20:28  adridg
-// Made exec() pure virtual for SyncActions
-//
-// Revision 1.57  2001/09/24 19:46:17  adridg
-// Made exec() pure virtual for SyncActions, since that makes more sense than having an empty default action.
-//
-// Revision 1.56  2001/09/23 18:24:59  adridg
-// New syncing architecture
-//
-// Revision 1.55  2001/09/16 12:24:54  adridg
-// Added sensible subclasses of KPilotLink, some USB support added.
-//
-// Revision 1.54  2001/09/06 22:04:27  adridg
-// Enforce singleton-ness & retry pi_bind()
-//
-// Revision 1.53  2001/09/05 21:53:51  adridg
-// Major cleanup and architectural changes. New applications kpilotTest
-// and kpilotConfig are not installed by default but can be used to test
-// the codebase. Note that nothing else will actually compile right now.
-//
